arreglando fallo de la TLB: Cuando un std::vector<T*> crece, puede ocurrir:     - el vector necesita más capacidad     - realoca su buffer interno     - copia los punteros a la nueva memoria     - pero cualquier puntero que apuntaba a elementos del vector anterior queda INVALIDADO
    En mi caso:
    root contenía punteros a nodos
    cada nodo tenía children, también un vector de punteros
    al hacer resize() en cualquiera de esos vectores, todos los punteros a nodos hijos quedaban colgando
Esto generaba:
    corrupción del árbol TLB
    nodos que parecían desplazarse
    rutas que dejaban de existir
    traducciones inconsistentes

error muy común en estructuras jerárquicas basadas en vector<T*>.

La solución correcta: usar std::unique_ptr<TLBNode>

Cuando cambio a:
std::vector<std::unique_ptr<TLBNode>> children;
ocurrió lo siguiente:
    el vector puede realocarse
    pero los nodos viven en el heap, no dentro del vector
    por tanto, sus direcciones NO cambian jamás
    y todos los punteros internos siguen siendo válidos

Un TLB multinivel es, en esencia, un árbol disperso:
PT2 -> PT1 -> PT -> DATA.

Se a implementado la ejecuccion de la isntruccion push y pop.

+++ include/runtime/rflags.h
@@ -28,7 +28,7 @@ typedef struct RFlags_t {
             uint8_t DM: 1; // Modo distribuido
         } bits;
 
-        uint8_t raw;
+        uint64_t raw;
     };
 } RFlags_t;
